archive/vite-adk-mcp/server/requirements.txt
fastapi>=0.79.0
uvicorn[standard]>=0.18.0
python-dotenv>=1.0.0
google-adk>=0.1.0
